delete page added

+++ projects/views.py
@@ -5,13 +5,13 @@ from .forms import ProjectForm
 def projects(request):
     projects = Project.objects.all()
     context = {'projects': projects}
-    return render(request, 'projects.html', context)
+    return render(request, 'projects/projects.html', context)
 
 def project(request, pk):
     project = Project.objects.get(id=pk)
     tags = project.tags.all()
     context = {'project': project, 'tags': tags}
-    return render(request, 'single-project.html', context)
+    return render(request, 'projects/single-project.html', context)
 
 def createProject(request):
     if request.method == 'POST':
@@ -22,7 +22,7 @@ def createProject(request):
 
     form = ProjectForm()
     context = {'form': form}
-    return render(request, 'project_form.html', context)
+    return render(request, 'projects/project_form.html', context)
 
 def updateProject(request, pk):
     project = Project.objects.get(id=pk)
@@ -35,9 +35,12 @@ def updateProject(request, pk):
 
     form = ProjectForm(instance = project)
     context = {'form': form}
-    return render(request, 'project_form.html', context)
+    return render(request, 'projects/project_form.html', context)
 
 def deleteProject(request, pk):
     project = Project.objects.get(id=pk)
-    project.delete()
-    return redirect('projects')
+    if request.method == 'POST':
+        project.delete()
+        return redirect('projects')
+    context = {}
+    return render(request, 'projects/delete_obj.html', context)
